+++ src/components/createLink.jsx
@@ -6,7 +6,6 @@ import { Link, useStaticQuery, graphql } from 'gatsby';
 
 import blogSlugs from '../../data/blog-slugs.json';
 import markdownPageSlugs from '../../data/markdown-page-slugs.json';
-import allFiles from '../__generated__/all-pages';
 import pageSlugs from '../data/page-slugs.json';
 
 const pageKeySlashIndex = 1;
@@ -37,7 +36,7 @@ const createLocalizedSlug = (locale, slug) => {
 
   const localeVersion = pageKey === '' ? `/${locale}` : `/${locale}${localized}`;
 
-  if (locale !== defaultLocale && allFiles.includes(localeVersion)) {
+  if (locale !== defaultLocale) {
     localized = localeVersion;
   }
 
@@ -57,11 +56,7 @@ export default (currentLocale) => {
 
       const linkPath = `${page}/${blogSlugs[filename][currentLocale]}`;
 
-      if (
-        filename in blogSlugs &&
-        currentLocale in blogSlugs[filename] &&
-        allFiles.includes(linkPath)
-      ) {
+      if (filename in blogSlugs && currentLocale in blogSlugs[filename]) {
         to = linkPath;
       }